- #28622 Adding aspects with default templates to async methods WIP
- #28984 Adding aspects with default templates to async iterator methods
- #28226 Test framework: generate/compare program test output.

+++ Tests/Caravela.Framework.Tests.InternalPipeline/Tests/Templating/NamespaceExpansion/Alias.t.cs
@@ -1,5 +1,5 @@
 int Method(int a)
 {
     global::System.Console.Write(global::System.Math.PI);
-    return (int)this.Method(a);
+    return this.Method(a);
 }
