+++ Tests/Caravela.Framework.Tests.InternalPipeline/Tests/Templating/Syntax/Switch/CompileTimeOldSwitch.t.cs
@@ -1,5 +1,5 @@
 int Method(int a)
 {
     global::System.Console.WriteLine("1");
-    return (int)this.Method(a);
+    return this.Method(a);
 }
